+++ simple_history/registry_tests/migration_test_app/models.py
@@ -13,3 +13,31 @@ class WhatIMean(DoYouKnow):
 class Yar(models.Model):
     what = models.ForeignKey(WhatIMean, on_delete=models.CASCADE)
     history = HistoricalRecords()
+
+
+class CustomAttrNameForeignKey(models.ForeignKey):
+    def __init__(self, *args, **kwargs):
+        self.attr_name = kwargs.pop("attr_name", None)
+        super(CustomAttrNameForeignKey, self).__init__(*args, **kwargs)
+
+    def get_attname(self):
+        return self.attr_name or super(
+            CustomAttrNameForeignKey, self
+        ).get_attname()
+
+    def deconstruct(self):
+        name, path, args, kwargs = super(
+            CustomAttrNameForeignKey, self
+        ).deconstruct()
+        if self.attr_name:
+            kwargs["attr_name"] = self.attr_name
+        return name, path, args, kwargs
+
+
+class ModelWithCustomAttrForeignKey(models.Model):
+    what_i_mean = CustomAttrNameForeignKey(
+        WhatIMean,
+        models.CASCADE,
+        attr_name='custom_attr_name',
+    )
+    history = HistoricalRecords()